Refactor GiantBomb module: rename file to `giant_bomb.rs`, update references, and remove old implementation
