+++ crates/providers/src/giant_bomb.rs
