+++ crates/providers/src/lib.rs
@@ -1,6 +1,6 @@
 pub mod anilist;
 pub mod audible;
-pub mod giantbomb;
+pub mod giant_bomb;
 pub mod google_books;
 pub mod hardcover;
 pub mod igdb;
